refactor analytics folder to contain logs and statistics
