+++ frontend/src/pages/analytics/logpages/PurchaseLogs/index.tsx
