+++ frontend/src/pages/analytics/logpages/TradeLogs/TradeDetail/TradeDetail.module.scss
