+++ frontend/src/pages/analytics/logpages/TradeLogs/TradeDetail/index.tsx
