+++ frontend/src/pages/analytics/logpages/TradeLogs/TradeLogs.module.scss
