+++ frontend/src/pages/analytics/logpages/TradeLogs/index.tsx
