+++ frontend/src/routes.tsx
@@ -15,9 +15,9 @@ import Stands from "./pages/admin/Stands";
 import Tags from "./pages/admin/Tags";
 import Volunteers from "./pages/admin/Volunteers";
 import LogHeader from "./components/pagePieces/LogsHeader";
-import TradesLogs from "./pages/logpages/TradeLogs";
-import PurchaseLogs from "./pages/logpages/PurchaseLogs";
-import Transactions from "./pages/analytics/Statistics/Transactions";
+import TradesLogs from "./pages/analytics/logpages/TradeLogs";
+import PurchaseLogs from "./pages/analytics/logpages/PurchaseLogs";
+import Transactions from "./pages/workspace/Transactions";
 import Statistics from "./pages/analytics/Statistics";
 import AuthPage from "./components/pagePieces/AuthBackground";
 import NotificationManager from "./components/NotificationManager";
